+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 **Closed issues:**
 
+- Fallback to FakeRedis if Redis is not available [\#412](https://github.com/EMMC-ASBL/oteapi-services/issues/412)
 - Revert update of codecov-action from v4 to v3 [\#325](https://github.com/EMMC-ASBL/oteapi-services/issues/325)
 - Add oteapi-dlite-plugin python package as a dependency [\#300](https://github.com/EMMC-ASBL/oteapi-services/issues/300)
 - Remove sftp service from production compose file [\#120](https://github.com/EMMC-ASBL/oteapi-services/issues/120)
@@ -22,7 +23,7 @@
 
 **Merged pull requests:**
 
-- \[Auto-generated\] Update dependencies [\#411](https://github.com/EMMC-ASBL/oteapi-services/pull/411) ([TEAM4-0](https://github.com/TEAM4-0))
+- Implement fallback to fakeredis [\#413](https://github.com/EMMC-ASBL/oteapi-services/pull/413) ([CasperWA](https://github.com/CasperWA))
 
 ## [v1.20230324.302](https://github.com/EMMC-ASBL/oteapi-services/tree/v1.20230324.302) (2023-03-24)